refactor: DRY some docs logic
