+++ src/lib/pushDoc.js
@@ -117,3 +117,21 @@ module.exports = async function pushDoc(key, selectedVersion, dryRun, filepath, 
       throw err;
     });
 };
+
+/**
+ * Recursively grabs all files within a given directory
+ * (including subdirectories)
+ * @param {String} folderToSearch path to directory
+ * @returns {String[]} array of files
+ */
+module.exports.readdirRecursive = function readdirRecursive(folderToSearch) {
+  const filesInFolder = fs.readdirSync(folderToSearch, { withFileTypes: true });
+  const files = filesInFolder
+    .filter(fileHandle => fileHandle.isFile())
+    .map(fileHandle => path.join(folderToSearch, fileHandle.name));
+  const folders = filesInFolder.filter(fileHandle => fileHandle.isDirectory());
+  const subFiles = [].concat(
+    ...folders.map(fileHandle => readdirRecursive(path.join(folderToSearch, fileHandle.name)))
+  );
+  return [...files, ...subFiles];
+};
